Fix a bug in timestamp format
strftime('%c') returns current date. It could have a single digit date
or double digit date. When date is single digit a space is added before
the date. So you get two spaces, instead of one, after month.

Logpad relies on the timestamp format to decide whether to add timestamp
to file or not. If timestamp is not constant result will be variable.

This commit removes this ambiguity and adds a constant timestamp that
handles single and double digit dates without any hitch.